requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
# click==8.1.8
# colorama==0.4.6
contourpy==1.3.1
coverage==7.7.0
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
# fastapi-pagination==0.12.34
# fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jose==1.0.0
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.4
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==15.0.1